% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_domestic_migration.R
\name{adjust_domestic_migration}
\alias{adjust_domestic_migration}
\title{Calculate a residual between populations and convert to domestic migration
flows}
\usage{
adjust_domestic_migration(
  popn,
  target,
  dom_in,
  dom_out,
  col_aggregation = c("year", "gss_code"),
  col_popn,
  col_target,
  rows_to_constrain = TRUE
)
}
\arguments{
\item{popn}{A data frame containing a population.}

\item{target}{A data frame containing a population at a lower resolution to
be constrained to. Must share column names with \code{popn} to enable a
join.}

\item{dom_in}{A data frame containing domestic migration inflows (usually
from an earlier model step). It will be scaled to include additional
inflows. Must contain a column named \code{dom_in} with inflows.}

\item{dom_out}{A data frame containing domestic migration outflows (usually
from an earlier model step). It will be scaled to include additional
outflows. Must contain a column named \code{dom_out} with outflows.}

\item{col_aggregation}{Character vector. Names of columns to join \code{popn}
to \code{target}.}

\item{col_popn}{A string with the name of the column containing population
counts in \code{popn}}

\item{col_target}{A string with the name of the column containing population
counts in \code{target}}

\item{rows_to_constrain}{A logical vector limiting the rows of \code{popn}
where this function is applied. Must be all TRUE or FALSE in each grouping
level specified by \code{col_aggregation}.}
}
\value{
A list of two data frames. \code{dom_in} contains the input domestic
  inflows scaled to include the calculated positive residuals. \code{dom_out}
  contains the input domestic outflows scaled to include the
  calulatednegative residuals.
}
\description{
A method of reconciling the differnece between a population and a target
constraint. Rather than scaling the whole population to match, the difference
can be used to calculate additional domestic migration flows that are added
or removed from the population.
}
\details{
The function calculates the difference between a population and a (lower
resolution) target, and then returns updated domestic in- and outmigration
rates that include the residual.
}
