% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_gross_migration.R
\name{adjust_gross_migration}
\alias{adjust_gross_migration}
\title{Calculate a residual between populations and convert to gross migration flows}
\usage{
adjust_gross_migration(
  popn,
  target,
  inflow_df,
  outflow_df,
  col_aggregation = c("year", "gss_code"),
  col_popn,
  col_target,
  col_inflow = "dom_in",
  col_outflow = "dom_out",
  rows_to_constrain = TRUE
)
}
\arguments{
\item{popn}{A data frame containing a population.}

\item{target}{A data frame containing a population to be constrained to.
Must share column names with \code{popn} to enable a join.}

\item{inflow_df}{A data frame containing migration inflows.
It will be scaled to include additional inflows.}

\item{outflow_df}{A data frame containing migration outflows.
It will be scaled to include additional outflows.}

\item{col_aggregation}{Character vector. Names of columns to join \code{popn}
to \code{target}.}

\item{col_popn}{A string with the name of the column containing population
counts in \code{popn}}

\item{col_target}{A string with the name of the column containing population
counts in \code{target}}

\item{col_inflow}{A strings with the name of the column containing inflow
counts in \code{inflow_df}. Default 'dom_in'}

\item{col_outflow}{A strings with the name of the column containing inflow
counts in \code{outflow_df}. Default 'dom_out'}

\item{rows_to_constrain}{A logical vector limiting the rows of \code{popn}
where this function is applied. Must be all TRUE or FALSE in each grouping
level specified by \code{col_aggregation}.}
}
\value{
A list of two data frames. \code{inflow} contains the input domestic
  inflows scaled to include the calculated positive residuals. \code{outflow}
  contains the input domestic outflows scaled to include the
  calculated negative residuals.
}
\description{
A method of reconciling the difference between a population and a target
constraint. Rather than scaling the whole population to match, the difference
is used to adjust gross migration flows.
}
\details{
The function calculates the difference between a population and a (lower
resolution) target, and then returns updated gross in- and out-migration
flows that include the residual.
}
