% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_on_sya.R
\name{age_on_sya}
\alias{age_on_sya}
\title{Age on a population}
\usage{
age_on_sya(popn)
}
\arguments{
\item{popn}{A data frame containing a cohort population, with year and age columns}
}
\value{
A data frame with age and year columns advanced by one year
}
\description{
Age on a population
}
