% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_city_wards.R
\name{aggregate_city_wards}
\alias{aggregate_city_wards}
\title{Aggregate City of London ward data up to borough level}
\usage{
aggregate_city_wards(df, data_col)
}
\arguments{
\item{df}{A dataframe containing ward-level data}

\item{data_col}{String. The column name of containing data in \code{df}}
}
\value{
A data frame with City of London wards aggregated to a single
  geographic aggregation
}
\description{
Sum data for individual wards in City of London up to the local authority
level. Return the dataframe with all other areas unchanged.
}
