% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_regions.R
\name{aggregate_regions}
\alias{aggregate_regions}
\title{Aggregate district-level data to regional totals}
\usage{
aggregate_regions(
  x,
  gss_col = "gss_code",
  col_aggregation = c("year", "age", "sex"),
  england = FALSE,
  lookup = "input_data/lookup/district_to_region.rds"
)
}
\arguments{
\item{x}{A data frame containing population/component data. The last column
is assumed to contain the data.}

\item{gss_col}{String with the column name to join to the regional lookup's
\code{gss_code} column. Default \code{"gss_code".}}

\item{col_aggregation}{A string of column names (other than geography) to
group by when aggregating to region. Default \code{c("year", "age",
"sex")}.}

\item{england}{Logical indicating whether or not an England total should also
be calculated. Default \code{FALSE}}

\item{lookup}{String. File path for a gss_code to region gss code lookup}
}
\value{
The input data frame with additional rows for English regions and
  Wales
}
\description{
Aggregate district-level data to regional totals
}
