% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_constraint.R
\name{apply_constraint}
\alias{apply_constraint}
\title{Apply a constraint to a dataframe.}
\usage{
apply_constraint(
  x,
  constraint,
  constraint_lookup,
  mapping,
  data_col = NULL,
  col_geog = "area_code"
)
}
\arguments{
\item{x}{dataframe. Data to be constrained}

\item{constraint}{dataframe. Constraint data}

\item{constraint_lookup}{dataframe. Lookup between operational geography and
constraint geography}

\item{mapping}{Character. Column aggregation to use in constraining}

\item{data_col}{String. Column in x containing data. Default NULL will use
last column in x.}

\item{col_geog}{String. Column in x containing operational geography}
}
\value{
A dataframe containing constrained values for the selected geography
}
\description{
Constrain values in one dataframe to those in another at a higher geography.#'
}
