% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births_constrain.R
\name{births_constrain}
\alias{births_constrain}
\title{Scale one population to match the totals of another}
\usage{
births_constrain(births, constraint)
}
\arguments{
\item{births}{A data frame containing births by mother's age data.}

\item{constraint}{A data frame containing births data at the same
resolution or lower.}
}
\value{
A data frame of scaled births by mother's year of age.
}
\description{
A wrapper for \code{get_scaling_factor} which prepares births data
to feed into the function and which appliued the function's output
back to the bith data resulting in sacled births.
}
\details{
The NPP births constraint is provided for ages 15-46 while the
model produces births for ages 15-49. This function wrangles the
input birth data so that it can be scaled using the \code{get_scaling_factor}
function.
}
