% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births_from_popn_fert.R
\name{births_from_popn_fert}
\alias{births_from_popn_fert}
\title{Apply fertility rates to a population.}
\usage{
births_from_popn_fert(popn, fertility, col_aggregation = c("year",
  "gss_code", "age", "sex"), col_age = "age", col_sex = "sex",
  col_count = "count", col_rate = "rate", col_births = "births",
  birthratio_m2f = 1.05)
}
\arguments{
\item{popn}{A data frame containing population data.}

\item{fertility}{A data frame containing fertiity rates per person per year.
If \code{popn} contains sex data specified by \code{col_sex} and this does
not, it will be assumed the rates are per woman (with a warning).}

\item{col_aggregation}{A character vector giving the names of columns to
which the output births will be aggregated. Each column must be present in
\code{popn}. The vector will be used to establish a mapping to any columns
that are also in \code{fertility}, so use a named vector if names differ,
e.g.\code{c("gss_code"="LSOA11CD")}. At least one column must be in
\code{fertility} to enable a join. Defaults to \code{c("year", "gss_code",
"sex", "age")}.}

\item{col_age}{String. Name for age column in inputs (if present) and output.
Default "age".}

\item{col_sex}{String. Name for sex column in inputs (if present) and output.
Set to NULL if you are not working with sex. Default "sex".}

\item{col_count}{String. Name for population count column in \code{popn}.
Default "count".}

\item{col_rate}{String. Name for fertility rate column (per person per year)
in \code{fertility}. Default "rate".}

\item{col_births}{String. Name for births column in output. Default "births".}

\item{birthratio_m2f}{Numeric. Ratio of male births to female births (if
outputting sex). Default 1.05.}
}
\value{
A data frame of births with one row for each distinct value of the
  \code{col_aggregation} columns of the input, plus a \code{col_births}
  column containing births, and a \code{col_age} column with value 0.
  All other input columns are discarded.
}
\description{
Given a cohort population and a data frame of fertility rates that can be
joined to the population, return a data table with the population's
aggregation levels, a births count split by a provided male:female births
ratio, and age set to zero.
}
\details{
Sex data (if used) must be provided as "male" and "female".
}
\examples{

library(births)

popn <- expand.grid(year = 2000, age=20:23, gss_code=c("a","b","c"), sex=c("female","male"), count = 100)
fert <- expand.grid(year = 2000, age=20:23, gss_code=c("a","b","c"), rate = 0.01)

pop_births <- births_from_popn_fert(popn,
                                    fert,
                                    colname_aggregation = c("year", "gss_code", "age", "sex"),
                                    col_age = "age",
                                    col_sex = "sex",
                                    col_count = "count",
                                    col_rate = "rate",
                                    col_births = "births",
                                    birthratio_m2f = 1.05)
# equivalent to
pop_births <- births_from_popn_fert(popn, fert)

}
