% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_mortality_rates.R
\name{calc_mean_rate}
\alias{calc_mean_rate}
\title{Calculate average rates based on past data}
\usage{
calc_mean_rate(rate_backseries, years_to_avg, last_data_year, rate_col)
}
\arguments{
\item{rate_backseries}{Dataframe. A set of rates by LA, sex and age}

\item{years_to_avg}{numeric. Number of years data to include in the average}

\item{last_data_year}{numeric. The final year of data to include in the average}

\item{rate_col}{Character. The name of the column containing the rates}
}
\value{
A data frame of mortality probabilities or fertility rates.
}
\description{
Given a backseries of rates the function will return an average rate
}
