% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_geomean_scaling_factors.R
\name{calculate_geomean_scaling_factors}
\alias{calculate_geomean_scaling_factors}
\title{Calculate a set of scaling factors to apply to fertility or mortality rates}
\usage{
calculate_geomean_scaling_factors(
  popn,
  future_rates,
  data_years,
  constraint,
  constraint_data_col
)
}
\arguments{
\item{popn}{A data frame containing a population at small area (ward or msoa)
by sex and age.}

\item{future_rates}{A data frame containing mortality or fertility rates at
local authority level by sex and age.}

\item{data_years}{Numeric vector of years to average.}

\item{constraint}{A dataframe of total births or deaths for each small area.}

\item{constraint_data_col}{A string giving the column with data in the
constraint file.}
}
\value{
A dataframe of scaling factors by sex and age for each small area.
}
\description{
For a backseries of years, given population for small areas and
fertility/mortality rates for a higher geography, calculate a set of
implied births/deaths. Compare these to observed births/deaths and
create scaling facrots for the rates. Finally, calculate the geometric
mean of the scaling factors.
}
\details{
The function calculates the difference between a population and a (lower
resolution) target, and then returns updated domestic in- and outmigration
rates that include the residual.
}
