% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_from_backseries.R
\name{calculate_mean_from_backseries}
\alias{calculate_mean_from_backseries}
\title{Calculate averagesbased on past data}
\usage{
calculate_mean_from_backseries(data_backseries, years_to_avg,
  last_data_year, data_col, col_aggregation = c("gss_code", "sex"))
}
\arguments{
\item{data_backseries}{Dataframe. A set of rates or flows by LA, sex and age}

\item{years_to_avg}{numeric. Number of years data to include in the average}

\item{last_data_year}{numeric. The final year of data to include in the average}

\item{data_col}{Character. The name of the column containing the data}

\item{col_aggregation}{Character. The columns to group_by}
}
\value{
A data frame of mortality probabilities, fertility rates, or
international migration flows/rates.
}
\description{
Given a backseries of data the function will return an average
}
