% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rate_by_regression.R
\name{calculate_rate_by_regression}
\alias{calculate_rate_by_regression}
\title{Use linear regression to calculate future values based on past data}
\usage{
calculate_rate_by_regression(data_backseries, years_to_avg, last_data_year,
  data_col, col_aggregation = c("gss_code", "sex"))
}
\arguments{
\item{data_backseries}{Dataframe. A set of values by LA, sex and age}

\item{years_to_avg}{numeric. Number of years data to include in the average}

\item{last_data_year}{numeric. The final year of data to include in the average}

\item{data_col}{Character. The name of the column containing the values}
}
\value{
A data frame of mortality probabilities, fertility rates, or
international migration flows/rates.
}
\description{
Given a backseries of values the function will return an future value
}
