% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_region_domestic_flows.R
\name{calculate_region_domestic_flows}
\alias{calculate_region_domestic_flows}
\title{Calculate domestic in, out and net migration for a group of LAs}
\usage{
calculate_region_domestic_flows(domestic_rates, nat_chng_popn,
  gss_code_list, region_name)
}
\arguments{
\item{domestic_rates}{A data frame containing a rates of domestic migration
between all LAs by age and sex}

\item{nat_chng_popn}{A data frame containing natutal change population by
sex and age}

\item{gss_code_list}{A character vector containing gss_codes}

\item{region_name}{A string giving the name of the region being output}
}
\value{
A data frame with domestic in, out and net flows by sex, age and year
  for the projection period
}
\description{
Calculate domestic in, out and net migration for a group of LAs
}
