% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_negative_values.R
\name{check_negative_values}
\alias{check_negative_values}
\title{Test a column in a dataframe for negatives and return a warning or amend the values}
\usage{
check_negative_values(x, data_col, warn_only = FALSE, change_value = 0)
}
\arguments{
\item{x}{A data frame.}

\item{data_col}{String. The name of a column in the data frame containing
numeric data.}

\item{warn_only}{Logical. When TRUE, only warn that the data_col contains
negatives. When change negative values to the change_value.}

\item{change_value}{Numeric. If \code{warn_only = FALSE} this the value that
negatives area changed to. Default 0}
}
\value{
The input data frame. If \code{warn_only} is FALSE, negative values
  of \code{data_col} are set to the \code{change_value} value.
}
\description{
If negatives exist print a warning. If the \code{warn_only} parameter is
FALSE then change the negatives to a user-defined value.
}
