% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_fertility.R
\name{complete_fertility}
\alias{complete_fertility}
\title{Complete the sex and age of a fertility dataframe}
\usage{
complete_fertility(
  fertility,
  population,
  col_sex_fert = "sex",
  col_sex_popn = col_sex_fert,
  col_age_fert = "age",
  col_age_popn = col_age_fert,
  col_rate = "rate"
)
}
\arguments{
\item{fertility}{A dataframe containing fertility data. Any age column must be numeric.}

\item{population}{A dataframe containing population data. Must contain sex/age if
fertility contains sex/age.  Any age column must be numeric.}

\item{col_sex_fert}{A character string with the fertility sex column name.
Takes NULL if there is no sex column. Must be NULL if \code{col_sex_popn} is NULL.}

\item{col_sex_popn}{A character string with the population sex column name.
Takes NULL if there is no sex column. Must be NULL if \code{col_sex_fert} is NULL.
Defaults to \code{col_sex_fert}.}

\item{col_age_fert}{A character string with the fertility age column name.
Takes NULL if there is no age column. Must be NULL if \code{col_age_popn} is NULL.}

\item{col_age_popn}{A character string with the population age column name.
Takes NULL if there is no age column. Must be NULL if \code{col_age_fert} is NULL.
Defaults to \code{col_age_fert}.}

\item{col_rate}{A character string with the fertility rate column name.}
}
\value{
A fertility dataframe that contains entries for all sexes present in the
  population dataframe, and age extended down to the lowest population age and up
  to the highest population age.  New entries are created with a rate value of 0.
}
\description{
#' This function has been deprecated. Use \code{complete_popn_dataframe}.
}
