% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_popn_dataframe.R
\name{complete_popn_dataframe}
\alias{complete_popn_dataframe}
\title{Complete the sex and age of a population or rates dataframe}
\usage{
complete_popn_dataframe(
  x,
  sexes = c("female", "male"),
  ages = 0:90,
  col_sex = "sex",
  col_age = "age",
  col_data = "rate",
  fill_value = 0
)
}
\arguments{
\item{x}{A dataframe containing population data. Any age column must be numeric.}

\item{sexes}{Character. The sexes to be imputed. Default c('female','male')}

\item{ages}{Numeric. The ages to be imputed. Default 0:90}

\item{col_sex}{String. The with the sex column name.
Takes NULL if there is no sex column.}

\item{col_age}{String. The the age column name.
Takes NULL if there is no age column.}

\item{col_data}{String. The with the data column name.}

\item{fill_value}{Numeric. The value to assign to the imputed levels.
Default 0.}
}
\value{
A dataframe that contains entries for all sexes in \code{sexes} and
all ages in \code{ages}. New entries are created with a rate value of \code{fill_value}.
}
\description{
Example use would be a fertility dataframe which only has data for
females aged 15-49. The function completes the dataframe for the entire
population with a rate of 0 for imputed aggregation levels..
}
