% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_from_popn_rate.R
\name{component_from_popn_rate}
\alias{component_from_popn_rate}
\title{Apply component rates to a population.}
\usage{
component_from_popn_rate(
  popn,
  component_rate,
  col_aggregation = c("year", "gss_code", "sex", "age"),
  col_popn = "popn",
  col_rate = "rate",
  col_component = "component"
)
}
\arguments{
\item{popn}{A data frame containing population data}

\item{component_rate}{A data frame containing component rates (per aggregation
level per year)}

\item{col_aggregation}{A string or character vector giving the names of
columns to which the output component will be aggregated to. If names differ
between input data frames, use a named character vector, e.g.
\code{c("popn_age"="mortality_age")}. Can contain columns that are not in
\code{component_rate}. Default \code{c("year", "gss_code", "age", "sex")}}

\item{col_popn}{String. Name of column in \code{popn} containing population
counts. Default "popn"}

\item{col_rate}{String. Name of column in \code{component_rate} containing
component rates. Default "rate"}

\item{col_component}{String. Name of column to write component to in the output.
Default "component"}
}
\value{
A data frame of the component with one row for each distinct value of the
  input \code{col_aggregation} column.
}
\description{
Given a cohort population and a data frame of component rates that can be
joined to the population, return a data table with the population's
aggregation levels and a component count.
}
\details{
If component would result in a negative population, an error will be thrown
(default) or the component will be capped at the population count.
}
\examples{

popn <- expand.grid(year=2000, age=20:21, gss_code=c("a","b"), sex=c("f","m"), popn = 100)
mortality <- expand.grid(year=2000, age=20:21, gss_code=c("a","b"), sex=c("f","m"), rate = 0.5)

deaths <- component_from_popn_rate(popn,
                          mortality,
                          col_aggregation = c("year", "gss_code", "sex", "age"),
                          col_popn = "popn",
                          col_rate = "rate",
                          col_component = "deaths")

# Due to default parameter values, this is equivalent to
deaths <- component_from_popn_rate(popn, mortality, col_component = "deaths")

}
