% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_proj_from_file.R
\name{component_proj_from_file}
\alias{component_proj_from_file}
\title{Read component projection from file (e.g. international in migration)}
\usage{
component_proj_from_file(filepath, first_proj_yr)
}
\arguments{
\item{filepath}{Filepath to the component RDS file}

\item{first_proj_yr}{First projection year}
}
\value{
A data frame of component projection
}
\description{
Given a filepath to a RDS file that contains component projection, return a
component projection dataframe
}
\examples{

int_in <- popmodules::component_proj_from_file(filepath = "my_int_in_file.rds", first_proj_yr = 2018)

}
