% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain_births.R
\name{constrain_births}
\alias{constrain_births}
\title{Scale births to match national totals by sex and age}
\usage{
constrain_births(births, constraint)
}
\arguments{
\item{births}{A data frame containing births by mother's age data.}

\item{constraint}{A data frame containing births data at the same
resolution or lower.}
}
\value{
A data frame of scaled births by mother's year of age.
}
\description{
A wrapper for \code{calculate_scaling_factors} which prepares births data to
feed into the function and then calculates scaled births.
}
\details{
Constraints are applied to sex, age and country (which is inferred from
gss_code), with each combination of these variables (where available)
constrained to the NPP national total.

The NPP births constraint is provided for ages 15-46 while the model produces
births for ages 15-49. This function wrangles the input birth to bin together
all births from ages higher than the NPP constraints in the highest bin. (It
does the same for ages lower, but throws an error if it sees any because
there shouldn't be any.)
}
