% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain_component.R
\name{constrain_component}
\alias{constrain_component}
\title{Scale one population to match the totals of another}
\usage{
constrain_component(
  popn,
  constraint,
  col_aggregation = c("year", "sex", "age", "country"),
  col_popn,
  col_constraint = col_popn
)
}
\arguments{
\item{popn}{A data frame containing population data.}

\item{constraint}{A data frame containing population data at the same
resolution or lower.}

\item{col_aggregation}{A string or character vector giving the join mapping between
\code{popn} and \code{constraint}. Equivalent to \code{by} in \code{dplyr} joins}

\item{col_popn}{String. Name of column in \code{popn} containing population
counts. Default "popn".}

\item{col_constraint}{String. Name of column in \code{constraint} containing population
counts. Default col_popn.}
}
\value{
A data frame of component counts calculated as input popn * rate, with one row
  for each distinct level of the input \code{popn} dataframe
}
\description{
Given a cohort population and a data frame of target marginal population
subtotals, return the input population scaled so that its population
subtotals match the target's at each grouping level.
}
\examples{

popn <- expand.grid(year=2000, age=20:21, gss_code=c("a","b"), sex=c("f","m"), popn = 100)
constraint <- expand.grid(year=2000, age=20:21, sex=c("f","m"), country = "a", popn = 400)

scaled <- constrain_component(popn,
                         constraint,
                         col_aggregation = c("year", "sex", "age", "country"),
                         col_popn = "popn",
                         col_constraint = "popn")

Due to default parameter values, this is equivalent to
scaled <- constrain_component(popn, constraint)

}
