% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain_cross_border.R
\name{constrain_cross_border}
\alias{constrain_cross_border}
\title{Scale a set of cross-border flows to match the totals of another}
\usage{
constrain_cross_border(
  domestic_flow,
  in_constraint,
  out_constraint,
  col_flow = "flow"
)
}
\arguments{
\item{domestic_flow}{A data frame containing origin-destination flows}

\item{in_constraint}{A data frame containing cross-border inflow totals at
the same resolution or lower.}

\item{out_constraint}{A data frame containing cross-border outflow totals at
the same resolution or lower.}

\item{col_flow}{String. Name of column in \code{domestic_flow} containing
population counts. Default "flow"}
}
\value{
A data frame of scaled domestic flows by authority, sex and single
  year of age.
}
\description{
A wrapper for \code{get_scaling_factor} which prepares origin-destination
flow data, constraining flows out of England to match the target data frame
in \code{in_constraint} and flows out of England to another the target data
frame in \code{out_constraint}.
}
