% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain_to_hma.R
\name{constrain_to_hma}
\alias{constrain_to_hma}
\title{Scale one population to match the totals of another in a housing market area}
\usage{
constrain_to_hma(
  popn,
  constraint,
  hma_list,
  col_aggregation = c("year", "hma", "sex", "age"),
  col_popn,
  col_constraint = col_popn
)
}
\arguments{
\item{popn}{A data frame containing population data.}

\item{constraint}{A data frame containing population data at the same
resolution or lower.}

\item{hma_list}{A named list where the name is an housing market area and the
elements are vectors of the gss_codes inside that HMA}

\item{col_aggregation}{A string or character vector giving the join mapping between
\code{popn} and \code{constraint}. Equivalent to \code{by} in \code{dplyr} joins}

\item{col_popn}{String. Name of column in \code{popn} containing population
counts. Default "popn".}

\item{col_constraint}{String. Name of column in \code{constraint} containing population
counts. Default col_popn.}
}
\value{
A data frame of component counts calculated as input popn * rate, with one row
  for each distinct level of the input \code{popn} dataframe
}
\description{
Given a cohort population and a data frame of target marginal population
subtotals, return the input population scaled so that its population
subtotals match the target's at each grouping level.
}
