% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_excels.R
\name{create_excels}
\alias{create_excels}
\title{Produce 3 Excel workbooks from a set of projection RDS files}
\usage{
create_excels(dir, stamp, file_name)
}
\arguments{
\item{dir}{String. The output directory where the projection RDS files are
saved.}

\item{stamp}{String. The time stamp on the output RDS file names format
\code{"\%y-\%m-\%d_\%H\%M"}.}

\item{file_name}{String. File name for the population workbook ouput. Must
include \code{.xslx} suffix.}
}
\description{
If a projection has been run with the \code{write_excel} variable set to
\code{FALSE} and the Excel outputs are subsequently required this function
will produce those workbooks. Creates a population workbook, an ONS household
workbook and a DCLG household workbook.
}
\examples{
create_excels(dir = "outputs/trend/2018/2018_long/",
  stamp = "19-11-13_2144",
  file_name = "2018_long_trend.xlsx")

}
