% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_border_constrain.R
\name{cross_border_constrain}
\alias{cross_border_constrain}
\title{Scale one population to match the totals of another}
\usage{
cross_border_constrain(domestic_flow, in_constraint, out_constraint,
  col_flow = "flow")
}
\arguments{
\item{domestic_flow}{A data frame containing origin-destination flows}

\item{in_constraint}{A data frame containing cross-border totals data at the same
resolution or lower.}

\item{out_constraint}{A data frame containing cross-border totals data at the same
resolution or lower.}

\item{col_flow}{String. Name of column in \code{domestic_flow} containing population
counts. Default "flow"}
}
\value{
A data frame of scaled domestic flows by authority, sex and single
  year of age.
}
\description{
A wrapper for \code{get_scaling_factor} which prepares origin-destination
flow data, constraining flows out of England to match the target data frame
in \code{in_constraint} and flows out of England to another the target data
frame in \code{out_constraint}.
}
