% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_border_constrain_all.R
\name{cross_border_constrain_all}
\alias{cross_border_constrain_all}
\title{Scale domestic outflows to matach a set of constraining flows}
\usage{
cross_border_constrain_all(domestic_flow, in_constraint,
  col_flow = "flow")
}
\arguments{
\item{domestic_flow}{A data frame containing origin-destination flows}

\item{in_constraint}{A data frame containing cross-border inflow totals data}

\item{col_flow}{String. Name of column in \code{domestic_flow} containing population
counts. Default "flow"}
}
\value{
A data frame of scaled domestic flows by authority, sex and
  single year of age.
}
\description{
A wrapper for \code{get_scaling_factor} which prepares origin-destination
flow data to feed into the function and which applies the function's output
back to the flow data resulting in sacled flows.
}
\details{
The function assumes all cross border flows in the system are scaled and
therefore only scales the inflows as this will, by definition, also scale
the outflows
}
