% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dclg_household_model.R
\name{dclg_stage_2}
\alias{dclg_stage_2}
\title{Implementation of the DCLG 2014 Stage 2 Household Model}
\usage{
dclg_stage_2(stage2_file_path, stage1_output)
}
\arguments{
\item{stage2_file_path}{String. File path to a dataframe containing
DCLG headship rates.}

\item{stage1_output}{A list of DCLG stage 1 model outputs}
}
\value{
A list containing 5 dataframes: unconstrained and constrained
  household projections, household and communal establishment populations,
  and households by age.
}
\description{
Produce household projections from an input population
projection using the 2014 DCLG household projection methodology
}
