% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deactivate_log.R
\name{deactivate_log}
\alias{deactivate_log}
\title{Write model log file}
\usage{
deactivate_log(log_file_path)
}
\arguments{
\item{log_file_path}{The path and file name (with extension) of the log file}
}
\description{
Run loggr::deactivate_log and then clean up the resulting log file
}
