% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths_null.R
\name{deaths_null}
\alias{deaths_null}
\title{Template deaths function that returns a fixed number of deaths (default zero)
for each aggregation level.}
\usage{
deaths_null(pop, col_aggregation = c("year", "gss_code", "age", "sex"),
  count = "value", const = 0, error_negative_pop = TRUE)
}
\arguments{
\item{pop}{A data frame containing population data}

\item{col_aggregation}{A string giving the names of columns to which the
output deaths will be aggregated to. Default \code{c("year", "gss_code",
"age", "sex")}}

\item{count}{String. Name of column with population counts. Default "value"}

\item{const}{Numeric. Number of deaths to return per geography. Defaults to
zero, but can be set to any positive number}

\item{error_negative_pop}{Logical. Throw error if deaths result in a negative
population count. Setting to FALSE creates a warning instead. Default TRUE.}
}
\value{
A data frame of deaths with one row for each distinct value of the
  \code{col_aggregation} columns, a column named deaths with value
  \code{const} and a column named age with value 0.
}
\description{
This is designed to be used as a placeholder deaths function during model
testing and a template for other deaths functions, along with its tests.
}
