% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute_within_age_band.R
\name{distribute_within_age_band}
\alias{distribute_within_age_band}
\title{Create a single year of age population}
\usage{
distribute_within_age_band(
  popn_1,
  popn_2,
  popn_1_col,
  popn_2_col,
  min_age,
  max_age,
  col_aggregation = c("gss_code", "sex"),
  additional_dist_cols = NULL
)
}
\arguments{
\item{popn_1}{A data frame containing population data for an age group.}

\item{popn_2}{A data frame containing a population at sya covering the same age group.}

\item{popn_1_col}{String. Name of column in \code{popn_1} containing the population data}

\item{popn_2_col}{String. Name of column in \code{popn_2} containing the population data}

\item{min_age}{Numeric. The minimum age of the age band.}

\item{max_age}{Numeric. The maximum age of the age band.}

\item{col_aggregation}{Character vector. The columns on which to join the two
datasets \code{popn_1} and \code{popn_2}. Cannot contain \code{"age"}.
Default \code{c("gss_code", "sex")}}

\item{additional_dist_cols}{Character vector. Columns that are in popn_2 but not in
popn_1 that you would like to retain in the output (e.g. "sex").}
}
\value{
A dataframe of population data by single-year-of-age for the input age range
}
\description{
Where a population is available for an age band only use the sya distribution
from another population to distribute the total over the single ages in the band.
}
