% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.add_slash}
\alias{.add_slash}
\title{#A function to check if a file path ends with a slash and add one if not}
\usage{
.add_slash(x)
}
\arguments{
\item{x}{A Character string}
}
\value{
The input string with a slash (`/`) on the end
}
\description{
#A function to check if a file path ends with a slash and add one if not
}
