% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.path_or_dataframe}
\alias{.path_or_dataframe}
\title{So that functions can either be passed the path to some data
or a dataframe}
\usage{
.path_or_dataframe(x)
}
\arguments{
\item{x}{A string or a dataframe}
}
\value{
dataframe
}
\description{
So that functions can either be passed the path to some data
or a dataframe
}
