% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.standardise_df}
\alias{.standardise_df}
\title{Select columns in a dataframe and rename the geography column to the generic
name 'area_code'}
\usage{
.standardise_df(
  x,
  col_geog,
  data_col = NULL,
  col_agg = c("year", "gss_code", col_geog, "age", "sex")
)
}
\arguments{
\item{x}{A dataframe}

\item{col_geog}{String. The geography column}

\item{data_col}{String. The data column. Default NULL}

\item{col_agg}{String. The column aggregation to select}
}
\value{
dataframe
}
\description{
Select columns in a dataframe and rename the geography column to the generic
name 'area_code'
}
