% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_fns_list.R
\name{evaluate_fns_list}
\alias{evaluate_fns_list}
\title{Evaluate a chain of functions passed as a list}
\usage{
evaluate_fns_list(fns_args_list)
}
\arguments{
\item{fns_args_list}{List of functions and their arguments}
}
\value{
The output from the final function in the funs_args_list
}
\description{
Each element of the arugment \code{fns_args_list} is a list.
These function lists must contain a function \code{fn} and
a list \code{args} of arguments. The output of one function will
be passed to the next as it's first argument (like a dplyr pipe).
}
\examples{
\dontrun{
 funs_args_list <- list(
  list(fn = function1, args = list(arg1_1, arg1_2, ....)),
  list(fn = function2, args = list(arg2_1, arg2_2, ....)),
  .....
)
}

}
