% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fert_from_file.R
\name{fert_from_file}
\alias{fert_from_file}
\title{Read fertility rates from file}
\usage{
fert_from_file(filepath)
}
\arguments{
\item{filepath}{Filepath to the fertility RDS file}
}
\value{
A data frame of fertility
}
\description{
Given a filepath to a RDS file that contains fertility rates, return a
fertility rate dataframe
}
\examples{

fert <- popmodules::fert_from_file(filepath = "my_fert_file.rds")

}
