% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_component_backseries.R
\name{get_component_backseries}
\alias{get_component_backseries}
\title{Divide a component of change by population to give historical rates}
\usage{
get_component_backseries(component_mye_path, popn_mye_path,
  births_mye_path, col_partial_match = NULL)
}
\arguments{
\item{component_mye_path}{Path to the mid-year estimates of the desired
component of change.}

\item{popn_mye_path}{Path to poplation mid-year estimates.}

\item{births_mye_path}{Path to births mid-year estimates.}

\item{col_partial_match}{List of columns in \code{component_mye_path} where
levels are a subset of those in \code{popn_mye_path}.}
}
\value{
A data frame containing rates for the component of change (in the
  "rate" column)
}
\description{
Takes filepaths to the mid year estimates for the desired component of
change, total population, and birth rates, and returns a data frame with
rates.
}
\details{
The function assumes input has the standard mid-year-estimate column names:
\code{c("gss_code","gss_name","geography","country","sex","age","year")} with
data stored in a \code{value} column.

If the components of change data is known to be at fewer levels than the
population mid-year estimates, then the \code{col_partial_match} parameter
should name the columns where this is the case. This will be true for example
with fertility data, which is limited to a certain age range.

Due to the way data comparisons work, the population must be agerd on as part
of the calculations, and the function also needs births estimates as input.
}
