% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_component.R
\name{get_component_from_file}
\alias{get_component_from_file}
\title{Read in a component data RDS file}
\usage{
get_component_from_file(filepath, max_yr)
}
\arguments{
\item{filepath}{String. A file path including the \code{.rds} suffix}

\item{max_yr}{Numeric. The maximum year of data required}
}
\value{
A data frame of component data
}
\description{
Read a RDS file containing population component data and filter
out years greater than a defined maximum year
}
