% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_constraints_from_file.R
\name{get_constraints_from_file}
\alias{get_constraints_from_file}
\title{Read in constraint files}
\usage{
get_constraints_from_file(popn_path, births_path, deaths_path, int_in_path,
  int_out_path, cross_in_path, cross_out_path)
}
\arguments{
\item{popn_path}{String}

\item{births_path}{String}

\item{deaths_path}{String}

\item{int_in_path}{String}

\item{int_out_path}{String}

\item{cross_in_path}{String}

\item{cross_out_path}{String}
}
\value{
List of dataframes
}
\description{
Read in constraint files
}
