% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gss_names.R
\name{get_gss_names}
\alias{get_gss_names}
\title{Return a data frame with local authority codes and names}
\usage{
get_gss_names()
}
\description{
Return a data frame with local authority codes and names
}
