% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mortality_backseries.R
\name{get_mortality_backseries}
\alias{get_mortality_backseries}
\title{Divide a component of change by population to give a data frame of rates}
\usage{
get_mortality_backseries(deaths_mye_path, popn_mye_path, births_mye_path)
}
\arguments{
\item{deaths_mye_path}{Path to deaths MYE}

\item{popn_mye_path}{Path to poplation MYE}

\item{births_mye_path}{Path to births MYE}
}
\value{
A data frame containing death rates (in the "value" column)
}
\description{
takes filepaths to the CoC and popn, and returns a rates df
}
