% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rates_flows_info.R
\name{get_rates_flows_info}
\alias{get_rates_flows_info}
\title{Convert an input list into a dataframe for use in the trend model}
\usage{
get_rates_flows_info(data_list, first_proj_yr, last_proj_yr)
}
\arguments{
\item{data_list}{A named list of lists. Each list name is a year. Each list
contains a list with a \code{path} and \code{transition} variable}

\item{first_proj_yr}{Numeric. The first projection year}

\item{last_proj_yr}{Numeric. The last projection year}
}
\value{
A dataframe containing paths and information for calculating rates
}
\description{
Information on rate or flows to be used in the trend model are passed to the
model as a named list. The list names are years and each element contains a
\code{path} variable and a \code{transition} variable.
}
\details{
\code{path} is a String. The file path to a dataframe containing rates or flows
\code{transition} is Logical. Is the year interpolating from one set of rates
  to another

The function extracts information from the list of lists and arranges it in a
dataframe which is used by the \code{get_rates_or_flows} function to read
and calculate model data
}
