% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rates_or_flows.R
\name{get_rates_or_flows}
\alias{get_rates_or_flows}
\title{Return a dataframe containing rates or flows for use in the trend model}
\usage{
get_rates_or_flows(
  df,
  df_info,
  projection_year,
  first_proj_yr,
  col_aggregation,
  data_col,
  geog_code_col,
  standardise_geog = TRUE
)
}
\arguments{
\item{df}{A data frame or NULL. The dataframe is the output of this function in
the previous year iteration}

\item{df_info}{A data frame. The output from the function \code{get_rates_flows_info}}

\item{projection_year}{Numeric}

\item{first_proj_yr}{Numeric. The first projection year}

\item{col_aggregation}{Character. The non-data columns in the dataframe \code{df}}

\item{data_col}{String. The column containing the flow or rate information to be
in the model}

\item{geog_code_col}{String. Which column in df contains the geography code}

\item{standardise_geog}{Logical. Should the geography column be renamed to the 
generic 'area_code' or left alone. flex model needs TRUE, trend model needs FALSE.
Default TRUE}
}
\value{
A dataframe containing rates or flows to use in the model and also
  rates or flows to be used in future calculations
}
\description{
A function to read in data and where necessary create transitional data between
one known data point and the next. The returned dataframe contains a data column
\code{data_col} with the rate or flow to use in the model and also columns \code{value_1}
and \code{value_2} to be used in the next call to this function.
}
