% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scaling_factors.R
\name{get_scaling_factors}
\alias{get_scaling_factors}
\title{Produce scaling factors to convert one population to another}
\usage{
get_scaling_factors(popn, constraint, col_aggregation = c("year", "sex",
  "age", "country"), col_popn, col_constraint = col_popn)
}
\arguments{
\item{popn}{A data frame containing population data.}

\item{constraint}{A data frame containing population data at the same
resolution or lower.}

\item{col_aggregation}{A string or character vector giving the join mapping between
\code{popn} and \code{constraint}. Equivalent to \code{by} in \code{dplyr} joins}

\item{col_popn}{String. Name of column in \code{popn} containing population
counts.}

\item{col_constraint}{String. Name of column in \code{constraint} containing population
counts. Defaults to col_popn.}
}
\value{
A data frame of scaling factors calculated as target population/input population,
  with one row for each level of the input population for each distinct level of the
  input \code{col_aggregation} columns.
}
\description{
Given a cohort population and a data frame of target marginal population
subtotals, return a set of scaling factors which, when applied, will scale
the input population to match the target at each grouping level.
}
