% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_model_outputs.R
\name{household_model_outputs}
\alias{household_model_outputs}
\title{Household model outputs}
\usage{
household_model_outputs(model_output, model, output_dir, timestamp,
  write_excel)
}
\description{
Household model outputs
}
