% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_mortality_rates.R
\name{initial_year_mortality_rate}
\alias{initial_year_mortality_rate}
\title{Calculate average mortality rates based on past data}
\usage{
initial_year_mortality_rate(mye_population, births, deaths,
  mortality_curves, last_data_year, years_to_avg, avg_or_trend)
}
\arguments{
\item{mye_population}{Dataframe. The mye backseries population.}

\item{births}{Dataframe. Births componet from the mye.}

\item{deaths}{Dataframe. Deaths component from the mye.}

\item{mortality_curves}{Dataframe. Age-and-sex-specific mortality
rate by local authority.}

\item{last_data_year}{Integer. The last year of death data on which to calculate
averages.}

\item{years_to_avg}{Integer. The number of years to use in calculating averages.}

\item{avg_or_trend}{Character. Should the averaged be caulated as the mean \code{Average},
or by linear regression \code{Trend}.}
}
\value{
A data frame of mortality probabilities.
}
\description{
Compares past mortality rates by LA, age and sex to an input mortality
curve to derive a set of scaling factors. The scaling factors are then
averaged or trended forward using regression. The resulting scaling
factor is applied to the input mortality curve to produce a set of
mortality rates.
}
