% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_mortality_rates.R
\name{initial_year_rate}
\alias{initial_year_rate}
\title{Calculate average rates based on past data}
\usage{
initial_year_rate(population, component_data, target_curves,
  last_data_year, years_to_avg, avg_or_trend, data_col, output_col)
}
\arguments{
\item{component_data}{String or data frame. If a string, an RDS filepath,
otherwise its contents. \code{births} or \code{deaths} by sex and age.}

\item{target_curves}{String or data frame. If a string, an RDS filepath,
otherwise its contents. Age-and-sex-specific mortality or fertiltiy
rates by local authority.}

\item{last_data_year}{numeric. The last year of death data on which to calculate
averages.}

\item{years_to_avg}{numeric. The number of years to use in calculating averages.}

\item{avg_or_trend}{Character. Should the averaged be caulated as the mean \code{Average},
or by linear regression \code{Trend}.}

\item{data_col}{Character. The column in the \code{component_data} dataframe containing the rates}

\item{output_col}{Character. The name of the column in the output dataframe containing the calculated rates}

\item{mye_population}{Character vector, data frame or list. If a character
vector, it contains filepaths to mye backseries population RDS file(s), otherwise the
contents of the file(s). For mortality rates the input is a list of 2
elements \code{population} and \code{births}. For fertility rates the
input is \code{population}}
}
\value{
A data frame of mortality probabilities or fertility rates by LA, year, sex and age.
}
\description{
Compares past rates by LA, age and sex to an input
curve to derive a set of scaling factors. The scaling factors are then
averaged or trended forward using regression. The resulting scaling
factor is applied to the input curve to produce a set of rates.
}
