% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_gla_models.R
\name{install_gla_models}
\alias{install_gla_models}
\title{Install the GLA model packages}
\usage{
install_gla_models(trend = TRUE, housing_led = TRUE, small_area = TRUE)
}
\arguments{
\item{trend}{Logical Document and install the trendmodel package Default \code{TRUE}}

\item{housing_led}{Logical Document and install the housingledmodel package Default \code{TRUE}}

\item{small_area}{Logical Document and install the smallareamodel package Default \code{TRUE}}
}
\description{
Run devtools::document() and devtools::install() on the GLA model packages for
first time installation or package updates. Specify which packages should be
updated using function parameters.
}
