% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/international_rates_and_flows.R
\name{international_rates_and_flows}
\alias{international_rates_and_flows}
\title{Calculate average international rates based on past data}
\usage{
international_rates_and_flows(popn_mye_path = NULL,
  births_mye_path = NULL, flow_or_rate, component_path, last_data_year,
  years_to_avg, data_col, n_proj_yr)
}
\arguments{
\item{popn_mye_path}{Character. The path to the mye backseries population.}

\item{component_path}{Character The path to the international in or out migration
componet from the mye.}

\item{last_data_year}{Integer. The last year of death data on which to calculate
averages.}

\item{years_to_avg}{Integer. The number of years to use in calculating averages.}

\item{data_col}{Character. The column in the component dataframe containing
the flow data.}

\item{n_proj_years}{Numeric. The number of years to project forward}
}
\value{
A data frame of international migration probabilities.
}
\description{
A wrepper for \code{calc_mean_rate} which prepares mye
and international in/out flow component data.
}
