% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gla_models.R
\name{load_gla_models}
\alias{load_gla_models}
\title{Load the development versions of the GLA model suite}
\usage{
load_gla_models(
  popmodules = TRUE,
  trend = TRUE,
  housing_led = TRUE,
  small_area = TRUE
)
}
\arguments{
\item{popmodules}{Logical Load the popmodules package Default \code{TRUE}}

\item{trend}{Logical Load the trendmodel package Default \code{TRUE}}

\item{housing_led}{Logical Load the housingledmodel package Default \code{TRUE}}

\item{small_area}{Logical Load the smallareamodel package Default \code{TRUE}}
}
\description{
Run devtools::load_all() on the GLA model packages and popmodules
}
