% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_warnings.R
\name{log_warnings}
\alias{log_warnings}
\title{Write warnings to a text file}
\usage{
log_warnings(file_path)
}
\arguments{
\item{file_path}{A string giving the directory path and the name
of the file to save.}
}
\description{
Write warnings to a text file
}
