% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_births.R
\name{null_births}
\alias{null_births}
\title{Template births function that returns a fixed number of births (default zero)
for each aggregation level, and sets their age to zero}
\usage{
null_births(pop, col_aggregation = c("year", "gss_code", "age", "sex"),
  const = 0, col_age = "age")
}
\arguments{
\item{pop}{A data frame containing population data}

\item{col_aggregation}{A string giving the names of columns to which the
output births will be aggregated to. Default \code{c("gss_code","sex",
"age")}}

\item{const}{Numeric. Number of births to return per geography. Defaults to
zero, but can be set to any positive number}

\item{col_age}{A string giving the name of the age column, if it exists in
the input, or if it needs to be created for the output. It needn't be an
aggregation level.}
}
\value{
A data frame of births with one row for each distinct value of the
  \code{col_aggregation} columns, a column named births with value
  \code{const} and a column named \code{col_age} with value 0.
}
\description{
This is designed to be used as a placeholder births function during model
testing and a template for other births functions, along with its tests.
}
\examples{

pop <- expand.grid(year = 2000, age=20:23, gss_code=c("a","b","c"), sex=c("f","m"), popn = 100)

pop_births <- null_births(pop,
                          colname_aggregation = c("year", "gss_code", "age", "sex"),
                          const = 0,
                          col_age = "age")

# equivalent to
pop_births <- null_births(pop)

}
