% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_household_model.R
\name{ons_household_model}
\alias{ons_household_model}
\title{Run the 2016 ONS household model}
\usage{
ons_household_model(population, stage1_file_path, stage2_file_path,
  communal_est_pop_path)
}
\arguments{
\item{population}{A data frame containing population data.}

\item{stage1_file_path}{String. Path to file containing DCLG stage 1
household inputs.}

\item{stage2_file_path}{String. Path to file containing DCLG stage 2
household inputs.}

\item{communal_est_pop_path}{String. Path to file containing communal
establishment population rates for the each year of the projection period.}
}
\value{
A list containing 2 lists: Stage 1 outputs and Stage 2 outputs.
}
\description{
Run stages 1 and 2 of the ONS household model and return outputs.
}
