% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_model_ons.R
\name{ons_stage_1}
\alias{ons_stage_1}
\title{Implementation of the ONS 2016 Household Model Stage 1}
\usage{
ons_stage_1(popn, hh_rep_rates_path, communal_est_pop_path, first_proj_yr)
}
\arguments{
\item{popn}{A data frame containing population data.}

\item{hh_rep_rates_path}{String. Path to file containing household
representative rates for the each year of the projection period.}

\item{communal_est_pop_path}{String. Path to file containing communal
establishment population rates for the each year of the projection period.}

\item{first_proj_yr}{Numeric. First year of the model projection. Elderly
communal populations are adjusted after this date.}
}
\value{
A list containing 4 dataframes: Unconstrained and constrained
  household projections, communal establishment and household populations for
  the projection period.
}
\description{
Produce household projections from an input population projection using the
2016 ONS household projection methodology.
}
