% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_household_model.R
\name{ons_stage_2}
\alias{ons_stage_2}
\title{Implementation of the ONS 2016 Stage 2 Household Model}
\usage{
ons_stage_2(stage_2_file_path, stage1_output)
}
\arguments{
\item{stage1_output}{A list of dataframes of ONS stage 1 model outputs.}

\item{stage2_file_path}{String. File path to a dataframe containing ONS
headship rates.}
}
\value{
A list containing 2 dataframes: unconstrained and constrained
  household projections.
}
\description{
Produce household projections from an input population projection using the
2016 ONS household projection methodology.
}
