% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_forward_flat.R
\name{project_forward_flat}
\alias{project_forward_flat}
\title{Project a population data frame forward by taking the most recent year and
repeating it for each year up to an input last_proj_yr}
\usage{
project_forward_flat(df, last_proj_yr)
}
\arguments{
\item{df}{A population data frame}

\item{last_proj_yr}{The last year of the projection}
}
\value{
A data frame containing the past years and the projection years
}
\description{
Project a population data frame forward by taking the most recent year and
repeating it for each year up to an input last_proj_yr
}
