% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_mortality_probs.R
\name{project_mortality_rates}
\alias{project_mortality_rates}
\title{Project mortality rates forward based on a national trend}
\usage{
project_mortality_rates(jump_off_rates, future_mortality, first_proj_yr,
  n_proj_yr, npp_var = "2018_principal")
}
\arguments{
\item{jump_off_rates}{Dataframe. A set of ASMRs for the first projection year.}

\item{future_mortality}{Dataframe. A national-level ASMR trajectory.}

\item{first_proj_yr}{Integer. The first projection year.}

\item{npp_var}{Character. Mortality trend variant. Either \code{2018_principal},
\code{2018_high}, \code{2018_low},  \code{2016_principal}, \code{2016_high}, or
\code{2016_low}. Defaults to \code{2018_principal}.}

\item{n_proj_year}{Integer. Number of years to project}
}
\value{
A data frame of mortality probabilities for all projection years.
}
\description{
Applies the national rate of change in mortality to a set of local
authority AMSRs to produce a mortality probability trajectory.
}
