% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_mortality_probs.R
\name{project_rates}
\alias{project_rates}
\title{Project mortality/fertility rates forward based on a national trend}
\usage{
project_rates(jump_off_rates, rate_col, rate_trajectory, first_proj_yr,
  n_proj_yr, npp_var = "2016_principal")
}
\arguments{
\item{jump_off_rates}{Dataframe. A set of LA/age/sex rates for the first projection year.}

\item{rate_col}{Character. The column in the jump_off_rates dataframe containing the rates.}

\item{rate_trajectory}{String or data frame. A national-level ASMR/ASFR
trajectory, or a filepath to it in RDS format}

\item{first_proj_yr}{Integer. The first projection year.}

\item{npp_var}{Character. NPP trend variant. Either \code{2018_principal},
\code{2018_high}, \code{2018_low},  \code{2016_principal}, \code{2016_high}, or
\code{2016_low}. Defaults to \code{2018_principal}.}

\item{n_proj_year}{Integer. Number of years to project}
}
\value{
A data frame of fertility probabilities for all projection years.
}
\description{
Applies the national rate of change in fertility to a set of local
authority ASMRs/ASFRs to produce a mortality/fertility probability trajectory.
}
