% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_rates_npp.R
\name{project_rates_npp}
\alias{project_rates_npp}
\title{Project mortality/fertility rates forward based on a national trend}
\usage{
project_rates_npp(jump_off_rates, rate_col, rate_trajectory_filepath,
  first_proj_yr, n_proj_yr, npp_var = "2018_principal")
}
\arguments{
\item{jump_off_rates}{Dataframe. A set of LA/age/sex rates for the first projection year
Likely calculated from \code{scaled_mortality_curve()} or \code{scaled_fertility_curve()}}

\item{rate_col}{Character. The column in the jump_off_rates dataframe containing the rates.}

\item{rate_trajectory_filepath}{Character. The filepath for the national-level ASMR/ASFR trajectory.}

\item{first_proj_yr}{Integer. The first projection year.}

\item{npp_var}{Character. NPP trend variant. Either \code{2018_principal},
\code{2018_high}, \code{2018_low},  \code{2016_principal}, \code{2016_high}, or
\code{2016_low}. Defaults to \code{2018_principal}.}

\item{n_proj_year}{Integer. Number of years to project}
}
\value{
A data frame of fertility probabilities for all projection years.
}
\description{
Applies the national rate of change in mortality/fertility to a set of local
authority ASMRs/ASFRs to produce a mortality/fertility probability trajectory.
This takes the projection for the first projection year as an input, and
extends it to the required number of projection years
}
