% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_from_file.R
\name{rates_from_file}
\alias{rates_from_file}
\title{Read component rates from file (e.g. fertility)}
\usage{
rates_from_file(filepath)
}
\arguments{
\item{filepath}{Filepath to the component RDS file}
}
\value{
A data frame of component rates
}
\description{
Given a filepath to a RDS file that contains component rates, return a
component rate dataframe
}
\examples{

fert <- popmodules::rates_from_file(filepath = "my_fert_file.rds")

}
