% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_gss_codes.R
\name{recode_gss_codes}
\alias{recode_gss_codes}
\title{Recode LA GSS Codes to a later geography}
\usage{
recode_gss_codes(
  df_in,
  col_geog = "gss_code",
  data_cols = last(names(df_in)),
  fun = list(sum),
  recode_to_year = 2013,
  aggregate_data = TRUE,
  recode_gla_codes = FALSE,
  code_changes_path = "input_data/lookup/district_changes_clean.rds"
)
}
\arguments{
\item{df_in}{A data frame containing gss_codes and data.}

\item{col_geog}{A string. The column which contains gss codes (defaults to
\code{gss_code}).}

\item{data_cols}{A string or character vector. The column(s) that contain the
data to be aggregated. Defaults to last column of input dataframe.}

\item{fun}{List. Function to be applied in aggregating data. Passed to the
.funs argument of \code{dplyr::summarise_all}. Default \code{list(sum)}.}

\item{recode_to_year}{Numeric. Conform to geography in which year. Default
\code{2013} (i.e. census geography).}

\item{aggregate_data}{Logical. If set to true multiple instances of the same
gss code will be aggregated using the function specified in \code{fun} parammeter.
Default to \code{TRUE}.}

\item{recode_gla_codes}{Logical. If set to \code{TRUE} recodes to the codes used in
2018-based and earlier projections. This is introduced for backward
compatibility and doesn't correspond to any single vintage year.
Default \code{FALSE}.
@param code_changes_path string. file path to the code changes database.}
}
\value{
The input dataframe with gss codes changed and data aggregated
}
\description{
Switches any gss codes to a later vintage
}
\details{
Aggregates data where rows with the same gss code are present

User must define the column containing the gss_code and the columns
which contain data. Data columns must be numeric or integer.
}
