% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_for_output.R
\name{reorder_for_output}
\alias{reorder_for_output}
\title{Reorder a dataframe by geographies for output}
\usage{
reorder_for_output(df, gss_col = "gss_code", order = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{gss_col}{A string containing the name of the column with gss codes.
Default "gss_code".}

\item{order}{A character string of the gss code geographic identifiers
(i.e. the first 3 characters of the gss code) to determine the order of
sorting. Default c("E92","N92","S92","W92","E12","S12","E06","E07","E08","E09","W06").}
}
\value{
The input data frame, reordered by the rankings in the order parameter
}
\description{
Sorts a data frame to put gss codes in a specified order. Codes are sorted
in the first instance by the type of geographic entity (eg. nation, region, district)
and then alphabetically within those groups.
}
