% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_fertility_curve.R
\name{scaled_fertility_curve}
\alias{scaled_fertility_curve}
\title{Calculate average fertility rates based on past data}
\usage{
scaled_fertility_curve(popn_mye_path, births_mye_path,
  target_curves_filepath, last_data_year, years_to_avg, avg_or_trend,
  data_col = "births", output_col)
}
\arguments{
\item{popn_mye_path}{Character. Path to the MYE population data}

\item{births_mye_path}{Character. Path to the MYE births component data}

\item{target_curves_filepath}{Character. Path to the SNPP target fertility curves}

\item{last_data_year}{numeric. The last year of death data on which to calculate
averages.}

\item{years_to_avg}{numeric. The number of years to use in calculating averages.}

\item{avg_or_trend}{Character. Should the averaged be caulated as the mean \code{Average},
or by linear regression \code{Trend}.}

\item{data_col}{Character. The column in the \code{births} dataframe containing the rates. Defaults to \code{births}}

\item{output_col}{Character. The name of the column in the output dataframe containing the calculated rates}
}
\value{
A data frame of mortality probabilities or fertility rates by LA, year, sex and age.
}
\description{
Compares past rates by LA, age and sex to an input
curve to derive a set of scaling factors. The scaling factors are then
averaged or trended forward using regression. The resulting scaling
factor is applied to the input curve to produce a set of rates.
}
