% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_mortality_curve.R
\name{scaled_mortality_curve}
\alias{scaled_mortality_curve}
\title{Calculate average mortality probabilities based on past data and assign to 
first projection year}
\usage{
scaled_mortality_curve(popn_mye_path, births_mye_path, deaths_mye_path,
  target_curves_filepath, last_data_year, n_years_to_avg, avg_or_trend,
  data_col = "deaths", output_col)
}
\arguments{
\item{popn_mye_path}{Character. Path to the MYE population data}

\item{births_mye_path}{Character. Path to the MYE births component data}

\item{deaths_mye_path}{Character. Path to the MYE deaths component data}

\item{target_curves_filepath}{Character. Path to the SNPP target mortality curves}

\item{last_data_year}{numeric. The last year of death and population data on which to calculate averages.}

\item{n_years_to_avg}{numeric. The number of years to use in calculating averages/trending forward.}

\item{avg_or_trend}{Character. Should the averaged be caulated as the mean \code{Average},
or by linear regression \code{Trend}.}

\item{data_col}{Character. The column in the \code{deaths} dataframe containing the rates. Defaults to \code{deaths}}

\item{output_col}{Character. The name of the column in the output dataframe containing the calculated rates}
}
\value{
A data frame of mortality probabilities rates by LA, year, sex and age with the same age structure
as the target curves and overall rates scaled so that they are consistent with past deaths.
}
\description{
Compares past rates by LA and sex to an input
curve to derive a set of scaling factors. The scaling factors are then
averaged or trended forward using regression. The resulting scaling
factor is applied to the age specific input curve to produce a set of rates.
}
