% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_births_and_split_by_sex_ratio.R
\name{sum_births_and_split_by_sex_ratio}
\alias{sum_births_and_split_by_sex_ratio}
\title{Sum a data frame with births by mother's age, and split babies by a sex ratio}
\usage{
sum_births_and_split_by_sex_ratio(
  births,
  birthratio_m2f = 1.05,
  geog_cols = "gss_code"
)
}
\description{
Sum a data frame with births by mother's age, and split babies by a sex ratio
}
