% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_domestic_flows.R
\name{sum_domestic_flows}
\alias{sum_domestic_flows}
\title{Calculate a residual between populations and convert to domestic migration
flows}
\usage{
sum_domestic_flows(domestic_flow, in_or_out, flow_col = "flow")
}
\arguments{
\item{domestic_flow}{A data frame containing sya/sex flows between geographic areas}

\item{in_or_out}{A string indicating whether the flows should be summed for
the destination (in) or origin (out) area.}

\item{flow_col}{A string. The name of the column containg the flow data. Defaul \code{flow}}
}
\value{
A dataframe with flows into or out of an area summed
}
\description{
A method of reconciling the differnece between a population and a target
constraint. Rather than scaling the whole population to match, the difference
can be used to calculate additional domestic migration flows that are added
or removed from the population.
}
\details{
Convert origin destiation flow data to gross flow by origin or destination
}
