% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_complete.R
\name{validate_complete}
\alias{validate_complete}
\title{Check that the number of rows in a dataframe is consistent with the expected
number of rows}
\usage{
validate_complete(
  x,
  test_cols = c("year", "gss_code", "gss_code_ward", "sex", "age"),
  nested_cols = c("gss_code", "gss_code_ward"),
  warn_only = FALSE
)
}
\arguments{
\item{x}{A dataframe}

\item{test_cols}{Character. The aggregation levels to be tested}

\item{nested_cols}{Character. Any of the aggregation levels in \code{completed_cols} 
which are nested. For example geography columns.}

\item{warn_only}{Logical. If TRUE the function will return a warning, if FALSE
an error is produced. Default FALSE.}
}
\description{
Note that this is a crude test and only checks that \code{nrow(x)} is consistent
with the expected number of rows. It says nothing about the content of those
rows. Failure of the test indicated incompleteness. Passing the test does
not indicate completeness but can be used as a crude proxy.
}
