% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config_list.R
\name{validate_config_list}
\alias{validate_config_list}
\title{Validate a model's config_list against an expected list of config settings}
\usage{
validate_config_list(config_list, expected_config)
}
\arguments{
\item{config_list}{List containing model config settings}

\item{expected_config}{Character vector containing the names of expected
elements in \code{config_list}}
}
\value{
TRUE, invisibly
}
\description{
The GLA population models all take a \code{config_list} as input, a list
containing every customisable model setting. This function compares the
config list against a provided vector giving the expected elements in the
config list. It warns if the list contains additional settings, and throws an
error if the list is missing something the model is expecting. Otherwise it
returns TRUE.
}
