% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model_config.R
\name{write_model_config}
\alias{write_model_config}
\title{Save a model config file as a .txt}
\usage{
write_model_config(config_list, output_dir = config_list$output_dir)
}
\arguments{
\item{config_list}{A list. The model config.}

\item{output_dir}{A string. The output path of the projection. Default \code{config_list$output_dir}}
}
\description{
Save model parameters in the model output directory for future reference
}
