% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rate_to_population.R
\name{apply_rate_to_population}
\alias{apply_rate_to_population}
\title{Apply rates to a population.}
\usage{
apply_rate_to_population(
  popn,
  rates,
  col_aggregation = c("year", "gss_code", "sex", "age"),
  col_popn = "popn",
  col_rate = "rate",
  col_out = "component",
  additional_popn_cols = NULL,
  additional_rate_cols = NULL,
  validate_geog = FALSE,
  col_geog = NULL
)
}
\arguments{
\item{popn}{A data frame containing population data}

\item{rates}{A data frame containing rates data}

\item{col_aggregation}{A string or character vector giving the names of
columns on which to join (\code{by} in \code{dplyr::left_join}).
If names differ between the two input data frames, use a named character
vector, e.g. \code{c("gss_code"="LSOA11CD")}. Default \code{c("year",
"gss_code", "age", "sex")}.}

\item{col_popn}{String. Name of column in \code{popn} containing population
counts. Default "popn".}

\item{col_rate}{String. Name of column in \code{rates} containing rate
data. Default "rate".}

\item{col_out}{String. Name of the resulting component count column (popn*rate)
in the output dataframe. Default "component".}

\item{additional_popn_cols}{String. Names of columns in \code{popn}
which should be included in the output but which are not included in
\code{col_aggregation}. Used when multiple aggregation levels apply to the
same rates. Default NULL.}

\item{additional_rate_cols}{String. Names of columns in \code{rates}
which should be included in the output but which are not included in
\code{col_aggregation}. Used when multiple rates apply to the same
aggregation level, e.g. out migration to multiple locations. The resulting
left join with \code{rates} can therefore return a much larger data
frame. Default NULL.}

\item{validate_geog}{Boolean. Should the geography in the two input dataframes
be checked using the \code{popmodules::validate_same_geog} function. Default FALSE.}

\item{col_geog}{String or NULL. The column containing geographic codes/names.
Must be set if \code{validate_geog = TRUE}. Column must be present in both
input dataframes. Default NULL.}
}
\value{
A data frame of component counts calculated as input popn * rate, with one row
  for each distinct level of the input \code{col_aggregation} columns.
}
\description{
Given a cohort population and a data frame of rates at the same or lower
resolution, return a data table of the population's aggregation levels, and a
component count resulting from the rate applied to the population.
}
\details{
The output dataframe will consist of the columns in the \code{popn} input
dataframe, plus any columns specified in \code{additional_rate_cols}, and the
output data column \code{col_out}. Output has 1 row for each distinct level of
the \code{col_aggregation}.
}
