% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_negative_values.R
\name{check_negative_values}
\alias{check_negative_values}
\title{Test a column in a dataframe for negatives, optionally setting to zero}
\usage{
check_negative_values(x, data_col, set_to_zero = TRUE, alt_value = NULL)
}
\arguments{
\item{x}{A data frame.}

\item{data_col}{String. The name of a column in the data frame containing
numeric data.}

\item{set_to_zero}{Logical. When TRUE, set negative values to zero. When
FALSE return the data frame as is.}

\item{alt_value}{Numeric or NULL. If `set_to_zero = TRUE` this parameter can
be used to set to a value other than zero}
}
\value{
The input data frame. If \code{set_to_zero} is TRUE, negative values
  of \code{data_col} are set to zero.
}
\description{
If negatives exist print a warning. If the \code{set_to_zero} parameter is
TRUE then change the negatives to 0.
}
