% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_popn_from_components.R
\name{construct_popn_from_components}
\alias{construct_popn_from_components}
\title{Add components to a starting population}
\usage{
construct_popn_from_components(
  start_population,
  addition_data,
  subtraction_data,
  col_aggregation = c("year", "gss_code", "sex", "age")
)
}
\arguments{
\item{start_population}{Data frame with initial population. All subsequent inputs should be at the same resolution.}

\item{addition_data}{a List of component dataframes contain data to be added to the start population}

\item{subtraction_data}{a List of component dataframes contain data to be subtracted from the start population}

\item{col_aggregation}{Aggregation columns common to all the above to join on.}
}
\value{
A data frame with all components added.
}
\description{
Shortcut function to combine population components
}
