% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_outputs.R
\name{datastore_outputs}
\alias{datastore_outputs}
\title{Datastore outputs}
\usage{
datastore_outputs(population, births, deaths, int_in, int_out, dom_in,
  dom_out, output_dir, file_name, output_date, write_excel)
}
\description{
Datastore outputs
}
