% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_model_outputs.R
\name{household_model_outputs}
\alias{household_model_outputs}
\title{Household model outputs}
\usage{
household_model_outputs(
  model_output,
  model,
  output_dir,
  timestamp,
  write_excel
)
}
\arguments{
\item{model_output}{The household model output, as created by
\code{household_model_ons} or \code{household_model_dclg}.}

\item{model}{String. Either "ons" or "dclg".}

\item{output_dir}{Location for output.}

\item{timestamp}{String containing timestamp to use in output filename.}

\item{write_excel}{Logical. Whether to output Excel .xls files.}
}
\description{
Household model outputs
}
