% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popn_age_on.R
\name{popn_age_on}
\alias{popn_age_on}
\title{Age on population}
\usage{
popn_age_on(
  popn,
  col_aggregation = c("year", "gss_code", "age", "sex"),
  col_age = "age",
  col_year = "year",
  col_data = "popn",
  timestep = 1,
  template_age_levels = NULL,
  births = NULL,
  col_births = "births",
  col_geog = "gss_code"
)
}
\arguments{
\item{popn}{Data frame containing a cohort population.}

\item{col_aggregation}{Character vector of columns in \code{popn} acting as
aggregation levels. Default \code{c("year","gss_code","age","sex")}. In the case
of multiple geography columns, only supply the highest resolution column;
other columns will be dropped.}

\item{col_age}{String denoting the age column. Default "age".}

\item{col_year}{String denoting the year column. Can be NULL. Default "year".}

\item{col_data}{String or Character denoting the column(s) containing data.
For multiple columns pass as a vector \code{c("popn","popn_2")}.
Default "popn".}

\item{timestep}{Numeric denoting the model time step in years. If age data
are numeric, they must be spaced by this interval. If they are an ordered
factor, the function will advance the age by one age band regardless of
this parameter. The \code{col_year} column, if present, will also be
increased by \code{timestep} years. Default 1.}

\item{template_age_levels}{Numeric vector or a factor giving the expected
levels for the population's ages. This is an optional validation tool that
runs an extra check to make sure all expected levels are present. Default
NULL.}

\item{births}{Data frame, 0 or \code{NULL}. Births data at the same
aggregation levels as \code{popn} but coded to the lowest age band that can
be joined to the output with \code{rbind}. If zero, then the output rows for
zero-year-olds, all of which are zero. If \code{NULL} then no
zero-year-olds are included in the output. Default \code{NULL}.}

\item{col_births}{String denoting the name of the column containing data in 
the \code{births} dataframe.}

\item{col_geog}{Character vector denoting the geography column(s). Default "gss_code".
Only used in validation. Only used when births dataframe is specified.}
}
\value{
A data frame containing the population with age advanced by one time
  step (or age band). The lowest age will not be present in this output. Rows
  are arranged by \code{col_aggregation}, in that order.
}
\description{
Given a cohort population, advance the cohort's age by one timestep/age band
and (if specified) increase the date year.
}
\details{
The function works with cohort age data as either numeric data or as an
ordered factor (e.g. age bands). If age is numeric then all age bands must be
equally spaced by the interval given in \code{timestep}, usually 1.

Ageing of the population is done by either adding the timestep to the age
(for a numeric age) or by incrementing to the next level of the factor (for
age as an ordered factor). If \code{col_year} is given, the year will also
be increased by \code{timestep}.

The highest age/age band is considered to be unbounded. This means that when
a population is aged on, the population in this band remains in this band,
and the age/band below it is added to it. An ageing population will all
eventually be moved into this age band if no other processes are operating.
}
