% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_forward_flat.R
\name{project_forward_flat}
\alias{project_forward_flat}
\title{Copy the hold year data into the projection years}
\usage{
project_forward_flat(df, first_proj_yr, n_proj_yr, hold_yr)
}
\arguments{
\item{df}{A data frame}

\item{first_proj_yr}{The first year of the projection}

\item{n_proj_yr}{The number of years being projected}

\item{hold_yr}{The hold year}
}
\value{
A data frame containing the past years and the projection years
}
\description{
Copy the hold year data into the projection years
}
