% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_gss_codes.R
\name{recode_gss_to_2011}
\alias{recode_gss_to_2011}
\title{Recode LA GSS Codes to 2011 geography}
\usage{
recode_gss_to_2011(df, col_geog = "gss_code", col_aggregation,
  fun = list(sum), aggregate_data = TRUE)
}
\arguments{
\item{df}{A data frame containing gss_codes and data.}

\item{col_geog}{A character vector. The column which contains gss codes
(defaults to \code{gss_code}).}

\item{col_aggregation}{A string or character vector giving column names for
levels of data aggregations: e.g. geographic area, age/age band, sex.}

\item{fun}{A list. A function to summarise the data. Must be provided in a
list. Defaults to \code{list(sum)}.}

\item{aggregate_data}{Logical. Whether to reaggregate the data after
recoding. If TRUE, all other columns will be combined using the new levels
and the function specified ny \code{fun}. If FALSE nothing will be done and
any newly duplicated levels will remain.}
}
\value{
The input dataframe with gss codes changed and data aggregated
}
\description{
Switches any non-2011 gss codes to Census 2011 gss codes Uses a hard-coded
list of changes
}
\details{
Aggregates data where rows with the same gss code are present

User must define the column containing the gss_code and the columns to group
by
}
