% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_for_output.R
\name{reorder_for_output}
\alias{reorder_for_output}
\title{Reorder a dataframe by geographies for output}
\usage{
reorder_for_output(
  df,
  gss_col = "gss_code",
  path_to_reorder = "input_data/lookup/output_order.rds",
  desc = FALSE
)
}
\arguments{
\item{df}{A data frame.}

\item{gss_col}{A string containing the name of the column with gss codes.
Default "gss_code".}

\item{path_to_reorder}{A string with a path to a file with data to reorder
by. Default \code{"input_data/lookup/output_order.rds"}. Must contain a
column matching \code{gss_col} and a column named \code{output_order}.}

\item{desc}{Logical. Reverse the order.}
}
\value{
The input data frame, reordered by the rankings in the file given by
  \code{path_to_reorder}.
}
\description{
Reorders a data frame to put national GSS codes first, followed by regional,
then local authorities. Uses the file in
\code{input_data/lookup/output_order.rds} to reorder by default.
}
