% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_fertility_curve.R
\name{scaled_fertility_curve}
\alias{scaled_fertility_curve}
\title{Calculate average fertility rates based on past data and assign to first
projection year}
\usage{
scaled_fertility_curve(
  popn,
  births,
  target_curves,
  last_data_year,
  n_years_to_avg,
  avg_or_trend,
  data_col = "births",
  output_col,
  project_rate_from = last_data_year + 1,
  col_geog = "gss_code"
)
}
\arguments{
\item{popn}{String or data frame. Path to population data, or a
data frame containing the data.}

\item{births}{String or data frame. Path to births data, or a 
data frame with the data.}

\item{target_curves}{String or data frame. Path to the SNPP
target fertility curves, or a data frame with the data.}

\item{last_data_year}{Numeric. The last year of births data on which to
calculate averages.}

\item{n_years_to_avg}{Numeric. The number of years to use in calculating
averages/trend forward.}

\item{avg_or_trend}{Character. Should the averaged be calculated as the mean
\code{Average}, or by linear regression \code{Trend}.}

\item{data_col}{Character. The column in the \code{births} dataframe
containing the rates. Defaults to \code{births}}

\item{output_col}{Character. The name of the column in the output dataframe
containing the calculated rates}

\item{project_rate_from}{Numeric. The year for which a rate is being
calculated. Default \code{last_data_year+1}}

\item{col_geog}{Character. The column containing geographic codes or names}
}
\value{
A data frame of fertility fertility rates by LA, year, sex and age
 with the same age structure as the target curves and overall rates scaled
 so that they are consistent with past births.
}
\description{
Compares past rates by LA, age and sex to an input
curve to derive a set of scaling factors. The scaling factors are then
averaged or trended forward using regression. The resulting scaling
factor is applied to the input curve to produce a set of rates.
}
