% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_mortality_curve.R
\name{scaled_mortality_curve}
\alias{scaled_mortality_curve}
\title{Calculate average mortality probabilities based on past data and assign to
first projection year}
\usage{
scaled_mortality_curve(
  popn,
  births,
  deaths,
  target_curves,
  last_data_year,
  n_years_to_avg,
  avg_or_trend,
  data_col = "deaths",
  output_col,
  project_rate_from = last_data_year + 1,
  col_aggregation = c("gss_code", "year", "sex"),
  col_geog = "gss_code"
)
}
\arguments{
\item{popn}{Character. Path to the MYE population data}

\item{births}{Character. Path to the MYE births component data}

\item{deaths}{Character. Path to the MYE deaths component data}

\item{target_curves}{Character. Path to the SNPP target mortality curves}

\item{last_data_year}{numeric. The last year of death and population data on
which to calculate averages.}

\item{n_years_to_avg}{numeric. The number of years to use in calculating
averages/trending forward.}

\item{avg_or_trend}{Character. Should the averaged be calculated as the mean
\code{Average}, or by linear regression \code{Trend}.}

\item{data_col}{Character. The column in the \code{deaths} dataframe containing
the rates. Defaults to \code{deaths}}

\item{output_col}{Character. The name of the column in the output dataframe
containing the calculated rates}

\item{project_rate_from}{Numeric. The year for which a rate is being calculated.
Default \code{last_data_year+1}.}
}
\value{
A data frame of mortality probabilities rates by LA, year, sex and age
  with the same age structure as the target curves and overall rates scaled so
  that they are consistent with past deaths.
}
\description{
Compares past rates by LA and sex to an input
curve to derive a set of scaling factors. The scaling factors are then
averaged or trended forward using regression. The resulting scaling
factor is applied to the age specific input curve to produce a set of rates.
}
