% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_complete.R
\name{validate_complete}
\alias{validate_complete}
\title{Check that all expected aggreation levels are present}
\usage{
validate_complete(
  x,
  col_aggregation = c("year", "gss_code", "gss_code_ward", "sex", "age"),
  col_geog = c("gss_code", "gss_code_ward"),
  additional_cols = NULL,
  warn_only = FALSE
)
}
\arguments{
\item{x}{A dataframe}

\item{col_aggregation}{Character. The aggregation levels to be tested}

\item{col_geog}{Charater. The column containing geography. These do not
have to be included in the col_aggregation parameter.}

\item{additional_cols}{Character. Columns not included in col_aggregation
but which need to be included in the test. Default NULL.}
}
\description{
Check that all expected aggreation levels are present
}
