% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_same_geography.R
\name{validate_same_geog}
\alias{validate_same_geog}
\title{Confirm two datasets have matching geography}
\usage{
validate_same_geog(df_1, df_2, col_1 = "gss_code", col_2 = "gss_code")
}
\arguments{
\item{df_1}{A dataframe containing a column of gss codes}

\item{df_2}{A dataframe containing a column of gss codes}

\item{col_1}{Character. The name of the column containing gss codes in df_1
(default 'gss_code')}

\item{col_2}{Character. The name of the column containing gss codes in df_2
(default 'gss_code')}
}
\value{
Warnings if there differences in the gss codes supplied
}
\description{
Performs a check that the unique set of gss_codes in one
dataframe is the same as the unique set of gss_codes in
another
}
